64-bit fix: X properties should be gulong, not guint: opacity setting fixed.
(Old svn revision: 27309)

+++ panel/panel-private.h
@@ -62,8 +62,8 @@ struct _PanelPrivate
     XfcePanelWidthType  full_width;
     gint                transparency;
 
-    guint               opacity;
-    guint               saved_opacity;
+    gulong              opacity;
+    gulong              saved_opacity;
 
     gint                block_autohide;
     gint                hide_timeout;
